requirements.txt
# Flask and related extensions
Flask==2.3.2
Flask-SQLAlchemy==3.1.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9

# Database driver for PostgreSQL
psycopg2-binary==2.9.6

# Authentication and security
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1

# CORS support
Flask-Cors==3.0.10

# Environment variable management
python-dotenv==1.0.0

# Testing tools
pytest==7.4.0
pytest-flask==1.2.0

# Linting and code quality
flake8==6.1.0
black==23.3.0
